fix: rename components dir, set max width for p
