+++ src/components/pages/contact-sales/hero/images/illustration-pricing.png
